fix: PTM.ALG.BOOLEANPROPERTY Rückgabetyp long→int + Property-Initialisierung
- XGlobal/GlobalImpl: UNO-IDL 'long' mappt auf Java 'int', nicht 'long' –
  der Typ-Mismatch verhinderte den korrekten UNO-Bridge-Aufruf
- EditierbaresZelleFormatHelper: schreibt 'editierbareFelderHervorheben=true'
  beim ersten Sheet-Aufbau ins Dokument, damit ptmbooleanproperty() nicht
  fallback-0 liefert solange der Nutzer die Einstellung nie geändert hat
- DocumentPropertiesHelper: initBooleanPropertyIfAbsent() als neue Hilfsmethode

+++ src/main/java/de/petanqueturniermanager/addin/XGlobal.java
@@ -96,5 +96,5 @@ public interface XGlobal extends XInterface {
 
     int ptmpouleanzdreiergruppen(int anzTeams);
 
-    long ptmbooleanproperty(String propname);
+    int ptmbooleanproperty(String propname);
 }
